requirements.txt
Django==4.2.24
django-htmx==1.24.1
django-allauth==65.11.2
django-bootstrap-v5==1.0.11
pillow==11.3.0
whitenoise==6.10.0
psycopg2-binary==2.9.10
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
openpyxl==3.1.5